+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `@rarimo/rarime`:
   - `RemoveCredentials` credentials RPC method
   - `ExportIdentity` RPC method
+  - `GetIdentity` RPC method
+
+### Changed
+- `@rarimo/rarime`:
+  - `CreateIdentity` - accept `privateKeyHex` optional param for identity creation
 
 ## [2.0.3] - 2024-02-06
 ### Changed